SSH Controls
Big update:
In listener.py created a "PICK" and "DROP" command. PICK used to pick up an object with one command. DROP used to drop off an object at specified move.

In the standalone SSH modified to send and receive commands via command file or manual input (for testing purposes).

SSHThread was modified to be an "automated" version of the standalone thread.  Opens commands.log and waits for the system to write commands.