finish moving coordinate conversion variable to config
The coordinate conversion variables have been changed so that rather than being constants, they are in a dictionary in the 'main' class. This allows them to be changed at runtime which is necessary if they are going to be read from a config file. And having them in a dictionary specifically allows them to be easily read from (and written to) the config file. A get_arm_coordinates() was also added to simplify the main loop.

Co-Authored-By: Micah <thoresonm1@gmail.com>